#1398: Per review: More code made guarded to remove indentation
Changed typo in name of function

Made delete_files return Boolean, print OS exception text in log

Moved functions from admin.py to models.py so they can be shared by models

Check on correct filename before deleting doubly linked file in normal delete
